separate responsibilities from `FrequencyAnalyzer`
