+++ libaudioviz/include/audioviz/fft/FrequencyAnalyzer.hpp
@@ -2,7 +2,6 @@
 
 #include <cmath>
 #include <functional>
-#include <tk-spline.hpp>
 #include <vector>
 #include <span>
 
@@ -13,33 +12,13 @@ namespace audioviz
 {
 
 /**
- * Analyzes wave data to produce a frequency spectrum. Allows further processing
- * of the resulting spectrum, such as scaling and interpolation.
+ * Analyzes wave data to produce a frequency spectrum using FFT.
+ * Focuses solely on FFT operations and amplitude computation.
+ * Use BinPacker and Interpolator classes for further spectrum processing.
  */
 class FrequencyAnalyzer
 {
 public:
-	enum class Scale
-	{
-		LINEAR,
-		LOG,
-		NTH_ROOT
-	};
-
-	enum class InterpolationType
-	{
-		NONE,
-		LINEAR,
-		CSPLINE,
-		CSPLINE_HERMITE
-	};
-
-	enum class AccumulationMethod
-	{
-		SUM,
-		MAX
-	};
-
 	using WindowFunction = std::function<float(int, int)>;
 
 	static const inline WindowFunction WF_HANNING = [](int i, int fft_size)
@@ -56,37 +35,13 @@ private:
 	int fft_size;
 	float inv_fft_size;
 
-	// nth root
-	int nth_root{2};
-	float nthroot_inv{1.f / nth_root};
-
 	fftwf_dft_r2c_1d fftw{fft_size};
-
-	// interpolation
-	tk::spline spline;
-	InterpolationType interp{InterpolationType::CSPLINE};
-
-	// output spectrum scale
-	Scale scale{Scale::LOG};
-
-	// method for accumulating amplitudes in frequency bins
-	AccumulationMethod am{AccumulationMethod::MAX};
 
 	// window function
 	WindowFunction window_func{WF_BLACKMAN};
 	int wf_i{3}; // for imgui
 
-	// struct to hold the "max"s used in `calc_index_ratio`
-	struct _scale_max
-	{
-		float linear, log, sqrt, cbrt, nthroot;
-		void calc(int new_max, float nthroot_inv);
-	} scale_max;
-
-	int bin_pack_input_size{};
-	std::vector<std::pair<int, int>> bin_pack_index_mapping;
 	std::vector<float, fftw_allocator<float>> window_values;
-	std::vector<double> m_spline_x, m_spline_y;
 
 public:
 	/**
@@ -106,40 +61,10 @@ public:
 	inline int get_fft_output_size() const { return fftw.output_size(); }
 
 	/**
-	 * Set interpolation type.
-	 * @param interp new interpolation type to use
-	 * @returns reference to self
-	 */
-	void set_interp_type(InterpolationType interp);
-
-	/**
 	 * Set window function.
-	 * @param interp new window function to use
-	 * @returns reference to self
+	 * @param wf new window function to use
 	 */
 	void set_window_func(WindowFunction wf);
-
-	/**
-	 * Set frequency bin accumulation method.
-	 * @param interp new accumulation method to use
-	 * @returns reference to self
-	 */
-	void set_accum_method(AccumulationMethod am);
-
-	/**
-	 * Set the spectrum's frequency scale.
-	 * @param scale new scale to use
-	 * @returns reference to self
-	 */
-	void set_scale(Scale scale);
-
-	/**
-	 * Set the nth-root to use when using the `NTH_ROOT` scale.
-	 * @param nth_root new nth_root to use
-	 * @returns reference to self
-	 * @throws `std::invalid_argument` if `nth_root` is zero
-	 */
-	void set_nth_root(int nth_root);
 
 #ifdef AUDIOVIZ_IMGUI
 	/**
@@ -165,12 +90,8 @@ public:
 
 	inline void execute_fft() const { fftw.execute(); }
 	void compute_amplitude(std::span<float> output);
-	void bin_pack(std::span<float> out, std::span<const float> in);
-	void interpolate(std::span<float> spectrum);
 
 private:
-	float calc_index_ratio(float i) const;
-	void compute_bin_pack_index_mappings(size_t out_size, size_t in_size);
 	void compute_window_values();
 };
